+++ backend/src/daoBean/PoslovniPartnerDaoLocal.java
@@ -3,6 +3,6 @@ package daoBean;
 import model.PoslovniPartner;
 import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDao;
 
-public interface PoslovniPartnerDaoLocal extends GenericDao<PoslovniPartner, Long>{
+public interface PoslovniPartnerDaoLocal extends GenericDao<PoslovniPartner, Integer>{
 
 }
